requirements.txt
opencv-python>=4.8.0
imagehash>=4.3.1
Pillow>=10.0.0
numpy>=1.24.0
tqdm>=4.65.0
scenedetect[opencv]>=0.6.3
yt-dlp>=2024.1.0
scikit-image>=0.21.0
flask>=3.0.0
watchdog>=3.0.0
psutil>=5.9.0

# Security note (VULN-11):
# After installing, pin exact versions for production:
#   pip install pip-audit pip-tools
#   pip-audit -r requirements.txt
#   pip-compile requirements.txt -o requirements.lock.txt
#   pip install -r requirements.lock.txt
